add config file

+++ src/mixCompressor.cpp
@@ -2,14 +2,13 @@
 #include "utils.h"
 #include "lossyComp.h"
 #include "losslessComp.h"
+#include "paqCompressor.h"
 #include "fileProcess.h"
 #include "task.h"
 #include <cstdio>
 #include <ctime>
 #include <iostream>
 #include <mutex>
-#include <condition_variable>
-
 /**
  * @description: 构造函数，根据参数进行相应的初始化
  * @param rel_err 最大相对误差
@@ -17,16 +16,17 @@
  * @param input 输入文件
  * @param mode 运行模式，压缩为COMPRESS，解压为DECOMPRESS
  */
-MixCompressor::MixCompressor(double rel_err, double avg_err, 
-	std::string input, char mode) : PW_REL_ERR_MAX(rel_err), AVG_ERR_MAX(avg_err), 
-	inputFilepath(input), mode(mode)
+MixCompressor::MixCompressor(CompressConf &cf) : PW_REL_ERR_MAX(cf.MAX_PW_REL_ERR), AVG_ERR_MAX(cf.MAX_AVG_ERR), 
+	inputFilepath(cf.paras.inputFile), mode(cf.paras.mode), blockSize(cf.blockSize)
 {
 	// lossyCompPtr = new LossyCompressor(rel_err, avg_err);
 	// losslessCompPtr = new LosslessCompressor();
-	lossyCompPtr = std::make_shared<LossyCompressor>(rel_err, avg_err);
+	lossyCompPtr = std::make_shared<LossyCompressor>(PW_REL_ERR_MAX, AVG_ERR_MAX);
 	losslessCompPtr = std::make_shared<LosslessCompressor>();
+	// paqCompPtr = std::make_shared<PaqCompressor>(inputFilepath, rel_err, avg_err);
 
-	blockSize = 3000;
+	// blockSize = 3000;
+	this->setConf(cf);
 	columnSize = 0;
 }
 
@@ -55,7 +55,7 @@ int MixCompressor::getFileLines(std::string inputFilepath)
 {
 	std::ifstream in(inputFilepath.c_str(), std::ios::in);
 	if (!in.is_open()) {
-		std::cout << "MixCompressor::getFileLines(), !in.is_open() " << inputFilepath << std::endl;
+		LOG(ERROR) << "MixCompressor::getFileLines(), !in.is_open() " << inputFilepath << std::endl;
 		return -1;
 	}
 	std::string line;
@@ -72,21 +72,20 @@ int MixCompressor::getFileLines(std::string inputFilepath)
  * @param
  * @return: 
  */
-int MixCompressor::compress_old()
+int MixCompressor::compress_ppmd()
 {
-	std::string suffix = inputFilepath.substr(inputFilepath.length()-4);
-	if (suffix.compare(".csv") != 0)
+	// paqCompPtr = std::make_shared<PaqCompressor>(inputFilepath, PW_REL_ERR_MAX, AVG_ERR_MAX);
+	if (!isFileSuffix(inputFilepath, ".csv"))
 	{
-		std::cout << "Don't support this type of file!!! Only CSV files!!!" << std::endl;
+		LOG(ERROR) << "Don't support this type of file!!! Only CSV files!!!" << std::endl;
 		exit(0);
 	}
-	
 	// get fileLines
-	std::cout << "inputFilepath = " << inputFilepath << std::endl;
+	LOG(INFO) << "compress_ppmd(), inputFilepath = " << inputFilepath << std::endl;
 	fileLines = getFileLines(inputFilepath);
 	if (fileLines < 0)
 	{
-		std::cout << "fileLines = " << fileLines << std::endl;
+		LOG(ERROR) << "fileLines = " << fileLines << std::endl;
 		exit(1);
 	}
 	
@@ -103,8 +102,14 @@ int MixCompressor::compress_old()
 	fileProcPtr = std::make_shared<FileProcessor>(&in, &tmp_out, blockSize);
 	fileProcPtr->setFileLines(fileLines);
 
-	columnSize = fileProcPtr->initWork_old();
-	std::cout << "columnSize = " << columnSize << std::endl;
+	std::string metadata;
+	std::string header;
+	std::string losslessAlgorithm = "ppmd";
+	// columnSize = fileProcPtr->initWork_old();
+	columnSize = fileProcPtr->initWork(losslessAlgorithm, metadata, header);
+	fileProcPtr->writeOneBlock2Tempfile(metadata);
+	fileProcPtr->writeOneBlock2Tempfile(header);
+	LOG(INFO) << "columnSize = " << columnSize << std::endl;
 
 	int block_count = 0;
 	int line_num_of_block;
@@ -156,12 +161,16 @@ int MixCompressor::compress_old()
 	in.close();
 	tmp_out.close();
 
-	std::cout << "lossless encoding, paq9a algorithm..." << std::endl;
-	losslessCompPtr->compressFile_paq9a(tempFilepath, outputFilepath);
-	std::cout << "finish compressFile_paq9a" << std::endl;
+	std::cout << "lossless encoding, ppmd algorithm..." << std::endl;
+	LOG(INFO) << "tempFilepath = " << tempFilepath << ", outputFilepath = " << outputFilepath << std::endl;
+	LosslessCompressor::compressFile_ppmd(tempFilepath, outputFilepath);
+	// losslessCompPtr->compressFile_ppmd("file/7.csv_0.50.tmp", "file/lxx-7.hw");
+	std::cout << "finish compressFile_ppmd" << std::endl;
 
-	deleteTmpFile(tempFilepath);
-
+	if (conf.remove_tmp_file)
+	{
+		deleteTmpFile(tempFilepath);
+	}
 	return 1;
 }
 
@@ -170,12 +179,13 @@ int MixCompressor::compress_old()
  * @param
  * @return: int 1表示正常
  */
-int MixCompressor::compress()
+int MixCompressor::compress_paq()
 {
-	std::string suffix = inputFilepath.substr(inputFilepath.length()-4);
-	if (suffix.compare(".csv") != 0)
+	paqCompPtr = std::make_shared<PaqCompressor>(inputFilepath, PW_REL_ERR_MAX, AVG_ERR_MAX);
+
+	if (!isFileSuffix(inputFilepath, ".csv"))
 	{
-		std::cout << "Don't support this type of file!!! Only CSV files!!!" << std::endl;
+		LOG(ERROR) << "Don't support this type of file!!! Only CSV files!!!" << std::endl;
 		exit(0);
 	}
 	
@@ -184,7 +194,7 @@ int MixCompressor::compress()
 	fileLines = getFileLines(inputFilepath);
 	if (fileLines < 0)
 	{
-		std::cout << "fileLines = " << fileLines << std::endl;
+		LOG(ERROR) << "fileLines = " << fileLines << std::endl;
 		exit(1);
 	}
 	
@@ -203,15 +213,15 @@ int MixCompressor::compress()
 
 	std::string metadatas;
 	std::string header;
-	columnSize = fileProcPtr->initWork(metadatas, header);
+	std::string losslessAlgorithm = "paq";
+	columnSize = fileProcPtr->initWork(losslessAlgorithm, metadatas, header);
 	std::cout << "columnSize = " << columnSize << std::endl;
 
 	int block_count = 0;
 	int line_num_of_block;
 
-	losslessCompPtr->compress_init_paq9a(tempFilepath, outputFilepath);
-	losslessCompPtr->compress_str_paq9a(metadatas);
-	losslessCompPtr->compress_str_paq9a(header);
+	paqCompPtr->compress_str_paq9a(metadatas);
+	paqCompPtr->compress_str_paq9a(header);
 
 	std::string lossless_str = "";
 
@@ -231,10 +241,10 @@ int MixCompressor::compress()
 			// TODO 目前是写入中间文件，改成直接无损压缩lossless_str（从内存压缩，减少io）
 			// fileProcPtr->writeOneBlock2Tempfile(lossless_str);
 			clock_t e = clock();
-			std::cout << "lossyCompPtr->compressOneBlock() time = " << (double)(e - s)/CLOCKS_PER_SEC << std::endl;
+			LOG(INFO) << "lossyCompPtr->compressOneBlock() time = " << (double)(e - s)/CLOCKS_PER_SEC << std::endl;
 			if (lossless_str.length() >= 1024*1024)
 			{
-				losslessCompPtr->compress_str_paq9a(lossless_str);
+				paqCompPtr->compress_str_paq9a(lossless_str);
 				{
 					std::string t;
 					t.swap(lossless_str);
@@ -267,23 +277,22 @@ int MixCompressor::compress()
 
 	if (lossless_str.length() > 0)
 	{
-		losslessCompPtr->compress_str_paq9a(lossless_str);
+		paqCompPtr->compress_str_paq9a(lossless_str);
 		{
 			std::string t;
 			t.swap(lossless_str);
 		}
 		lossless_str = "";
 	}
-	losslessCompPtr->compress_paq9a_end();
+	paqCompPtr->compress_paq9a_end();
 
 	in.close();
 	tmp_out.close();
 
-	// std::cout << "lossless encoding, paq9a algorithm..." << std::endl;
-	// losslessCompPtr->compressFile_paq9a(tempFilepath, outputFilepath);
-	// std::cout << "finish compressFile_paq9a" << std::endl;
-
-	deleteTmpFile(tempFilepath);
+	if (conf.remove_tmp_file)
+	{
+		deleteTmpFile(tempFilepath);
+	}
 
 	return 1;
 }
@@ -293,9 +302,15 @@ void MixCompressor::setTask(std::shared_ptr<Task> &t)
 	this->task = t;
 }
 
-int nready = 0;
-std::mutex mut;
-std::condition_variable cond;
+void MixCompressor::setConf(CompressConf &cf)
+{
+	this->conf = cf;
+}
+
+
+// int nready = 0;
+// std::mutex mut;
+// std::condition_variable cond;
 /**
  * @description: 多线程版本，使用paq算法
  * @param
@@ -303,28 +318,28 @@ std::condition_variable cond;
  */
 int MixCompressor::compress_thread()
 {
-	std::string suffix = inputFilepath.substr(inputFilepath.length()-4);
-	if (suffix.compare(".csv") != 0)
+	if (!isFileSuffix(inputFilepath, ".csv"))
 	{
-		std::cout << "Don't support this type of file!!! Only CSV files!!!" << std::endl;
+		LOG(ERROR) << "Don't support this type of file!!! Only CSV files!!!" << std::endl;
 		exit(0);
 	}
 	
+	std::string losslessAlgorithm = "paq";
+
 	// get fileLines
 	std::cout << "inputFilepath = " << inputFilepath << std::endl;
 	fileLines = getFileLines(inputFilepath);
 	if (fileLines < 0)
 	{
-		std::cout << "fileLines = " << fileLines << std::endl;
+		LOG(ERROR) << "fileLines = " << fileLines << std::endl;
 		exit(1);
 	}
 	
+	// 计算文件块数目
 	blocks = (fileLines - 1) / blockSize + (((fileLines - 1) % blockSize) != 0);
 
-	std::string avg_err_str;
-	convertDouble(AVG_ERR_MAX, avg_err_str);
-	tempFilepath = inputFilepath  + "_" + avg_err_str + ".tmp";
-	outputFilepath = inputFilepath + "_" + avg_err_str + ".hw";
+	// 根据误差和输入文件，设置中间文件名、输出文件名
+	setPaths(AVG_ERR_MAX, inputFilepath, tempFilepath, outputFilepath);
 
 	std::ifstream in(inputFilepath.c_str(), std::ios::in);
 	std::ofstream tmp_out(tempFilepath.c_str(), std::ios::out | std::ios::trunc);
@@ -334,25 +349,19 @@ int MixCompressor::compress_thread()
 
 	std::string metadatas;
 	std::string header;
-	columnSize = fileProcPtr->initWork(metadatas, header);
+	columnSize = fileProcPtr->initWork(losslessAlgorithm, metadatas, header); // get metadatas and header of input file
 	std::cout << "columnSize = " << columnSize << std::endl;
 
 	int block_count = 0;
 	int line_num_of_block;
 
-	
-	// losslessCompPtr->compress_init_paq9a(tempFilepath, outputFilepath);
-	// losslessCompPtr->compress_str_paq9a(metadatas);
-	// losslessCompPtr->compress_str_paq9a(header);
-	// 将metadata和header添加到待压缩队列中
-	// task->push_str(metadatas);
-	// task->push_str(header);
+	// 将metadata和header插入队列，待压缩
 	{
-		std::unique_lock<std::mutex> lck(mut);
+		std::unique_lock<std::mutex> lck(task->mut);
 		task->push_str(metadatas);
 		task->push_str(header);
-		nready += 2;
-		cond.notify_all();
+		task->nready += 2;
+		task->cond.notify_all();
 	}
 
 	std::string lossless_str = "";
@@ -363,36 +372,25 @@ int MixCompressor::compress_thread()
 
 		clock_t s = clock();
 		line_num_of_block = fileProcPtr->getOneBlock(block);
-		// clock_t e_b = clock();
-		// std::cout << "fileProcPtr->getOneBlock() time = " << (double)(e_b - s) / CLOCKS_PER_SEC << std::endl;
 		if (line_num_of_block > 0)
 		{
 			++block_count;
 			lossyCompPtr->compressOneBlock(block, line_num_of_block);
-			// clock_t e_lossy = clock();
-			// std::cout << "lossy compress time = " << (double)(e_lossy - e_b) / CLOCKS_PER_SEC << std::endl;
-			// std::string lossless_str;
 			losslessCompPtr->compressOneBlock(block, line_num_of_block, lossless_str);
 			clock_t e = clock();
-			std::cout << "losslessCompPtr->compressOneBlock time = " << (double)(e - s) / CLOCKS_PER_SEC << std::endl;
-			// if (lossless_str.length() >= 1 * 1024 * 1024)
-			// {
-				// losslessCompPtr->compress_str_paq9a(lossless_str);
+			LOG(INFO) << "losslessCompPtr->compressOneBlock time = " << (double)(e - s) / CLOCKS_PER_SEC << std::endl;
 				// 写入task的队列中，待paq压缩
 				{
-					std::unique_lock<std::mutex> lck(mut);
+					std::unique_lock<std::mutex> lck(task->mut);
 					task->push_str(lossless_str);
-					++nready;
-					cond.notify_all();
+					++task->nready;
+					task->cond.notify_all();
 				}
 				{
 					std::string t;
 					t.swap(lossless_str);
 				}
-				lossless_str = "";
-			// }
-
-			
+				lossless_str = "";	
 			std::cout << "\rlossy processing... " << std::fixed << (double)block_count / blocks * 100 << " %" << std::endl;
 			std::cout.flush();
 		}
@@ -401,6 +399,7 @@ int MixCompressor::compress_thread()
 			std::cout << "\nlossy process " << block_count << " blocks\n"<< std::endl;
 			break;
 		}
+		// 释放内存
 		for(std::vector<std::string> x : block)
 		{
 			for (std::string s : x)
@@ -415,9 +414,9 @@ int MixCompressor::compress_thread()
 		block.shrink_to_fit();
 	}
 
+	// 最后一块
 	if (lossless_str.length() > 0)
 	{
-		// losslessCompPtr->compress_str_paq9a(lossless_str);
 		task->push_str(lossless_str);
 		{
 			std::string t;
@@ -429,7 +428,11 @@ int MixCompressor::compress_thread()
 	in.close();
 	tmp_out.close();
 
-	deleteTmpFile(tempFilepath);
+	if (conf.remove_tmp_file)
+	{	
+		deleteTmpFile(tempFilepath);
+	}
+	
 
 	task->setFinish();
 
@@ -441,20 +444,9 @@ int MixCompressor::compress_thread()
  * @param 
  * @return: int 1表示正常 
  */
-int MixCompressor::decompress()
+int MixCompressor::decompress(std::string &losslessAlgorithm)
 {
-	// std::ifstream in(inputFilepath.c_str(), std::ios::in);
-
-	// lossless decompress
-	// get metadata
-	// loop
-	// 		decompress one block
-	// end loop
-	std::string suffix = inputFilepath.substr(inputFilepath.length()-3);
-	if (suffix.compare(".hw") == 0) {
-		// tempFilepath = inputFilepath.replace(inputFilepath.find(".hw"), 3, ".tmp");
-		// std::cout << "tempFilepath = " << tempFilepath << std::endl;
-		// outputFilepath = inputFilepath.replace(tempFilepath.find(".tmp"), 4, "");
+ 	if (isFileSuffix(inputFilepath, ".hw")) {
 		std::vector<std::string> filename;
 		splitString(inputFilepath, filename, ".hw");
 
@@ -463,22 +455,24 @@ int MixCompressor::decompress()
 		std::cout << "tempFilepath = " << tempFilepath << std::endl;
 		std::cout << "outputFilepath = " << outputFilepath << std::endl;
 	} else {
-		std::cout << "decompress: wrong file type" << std::endl;
-		exit(0);
+		LOG(ERROR) << "decompress: wrong file type" << std::endl;
+		exit(1);
+	}
+
+	if (losslessAlgorithm.compare("paq") == 0)
+	{
+		std::cout << "lossless decoding, paq9a algorithm" << std::endl;
+		// paqCompPtr->decompressFile_paq9a(inputFilepath, tempFilepath);
+		PaqCompressor::decompressFile_paq9a(inputFilepath, tempFilepath);
+		std::cout << "decompress file to " << tempFilepath << std::endl;
+	}
+	else if (losslessAlgorithm.compare("ppmd") == 0)
+	{
+		std::cout << "lossless decoding, ppmd algorithm" << std::endl;
+		LosslessCompressor::decompressFile_ppmd(inputFilepath, tempFilepath);
+		std::cout << "decompress file to " << tempFilepath << std::endl;
 	}
 	
-	// tempFilepath = inputFilepath + ".tmp";
-	// outputFilepath = inputFilepath + ".hw";
-
-	// losslessCompPtr->decompressFile_7z(inputFilepath);
-
-	// std::cout << "lossless decoding, PPMD algorithm" << std::endl;
-	// losslessCompPtr->decompressFile_ppmd(inputFilepath, tempFilepath);
-
-
-	std::cout << "lossless decoding, paq9a algorithm" << std::endl;
-	losslessCompPtr->decompressFile_paq9a(inputFilepath, tempFilepath);
-	std::cout << "decompress file to " << tempFilepath << std::endl;
 	
 	std::ifstream tmp_in(tempFilepath.c_str(), std::ios::in);
 	std::ofstream out(outputFilepath.c_str(), std::ios::out | std::ios::trunc);
@@ -487,10 +481,10 @@ int MixCompressor::decompress()
 	fileProcPtr = std::make_shared<FileProcessor>(&tmp_in, &out, blockSize);
 
 	fileProcPtr->getMetadata(blockSize, columnSize, fileLines, blocks);
-	std::cout << "blockSize = " << blockSize << std::endl;
-	std::cout << "columnSize = " << columnSize << std::endl;
-	std::cout << "fileLines = " << fileLines << std::endl;
-	std::cout << "blocks = " << blocks << std::endl;
+	LOG(INFO) << "blockSize = " << blockSize << std::endl;
+	LOG(INFO) << "columnSize = " << columnSize << std::endl;
+	LOG(INFO) << "fileLines = " << fileLines << std::endl;
+	LOG(INFO) << "blocks = " << blocks << std::endl;
 
 	fileProcPtr->writeHeader2DecompressedFile();
 	
@@ -518,7 +512,11 @@ int MixCompressor::decompress()
 	tmp_in.close();
 	out.close();
 
-	deleteTmpFile(tempFilepath);
+	if (conf.remove_tmp_file)
+	{
+		deleteTmpFile(tempFilepath);
+	}
+	
 	std::cout << "\nfinish decompress all " << block_count << " blocks" << std::endl;
 	std::cout << "decompress file " << inputFilepath << " to file " << outputFilepath << std::endl;
 	return 1;
@@ -534,15 +532,50 @@ void MixCompressor::run()
 	if (mode == COMPRESS)
 	{
 		clock_t startTime = clock();
-		// compress_old();
-		// compress();
-		compress_thread();
+		if (conf.multi_thread)
+		{
+			LOG(INFO) << "multi thread, compress_paq()" << std::endl;
+			compress_thread();
+		}
+		else
+		{
+			if (conf.lossless_algorithm.compare("paq") == 0)
+			{
+				LOG(INFO) << "single thread, compress_paq()" << std::endl;
+				compress_paq();
+			}
+			else if (conf.lossless_algorithm.compare("ppmd") == 0)
+			{
+				LOG(INFO) << "single thread, compress_ppmd()" << std::endl;
+				compress_ppmd();
+			}
+			else {
+				LOG(ERROR) << "don't support this lossless algorithm: " << conf.lossless_algorithm << std::endl;
+				exit(1);
+			}
+		}
 		clock_t endTime = clock();
 		std::cout << "MixCompressor::run() total time = " << (double)(endTime - startTime) / CLOCKS_PER_SEC << " seconds" << std::endl;
 	}
 	else if (mode == DECOMPRESS) {
 		clock_t startTime = clock();
-		decompress();
+		char buf[10] = {0};
+		sprintf(buf, "pQ9%c", 1);
+		std::string paq_version = buf;
+		if (FileProcessor::checkFile(inputFilepath, paq_version))
+		{
+			// paq
+			LOG(INFO) << "decompress paq file" << std::endl;
+			std::string lossless = "paq";
+			decompress(lossless);
+		}
+		else
+		{
+			// ppmd
+			LOG(INFO) << "decompress ppmd file" << std::endl;
+			std::string lossless = "ppmd";
+			decompress(lossless);
+		}
 		clock_t endTime = clock();
 		std::cout << "decompress total time = " << (double)(endTime - startTime) / CLOCKS_PER_SEC << " seconds" << std::endl;
 	}
@@ -559,7 +592,7 @@ int MixCompressor::deleteTmpFile(std::string &tmpfile)
 	{
 		return 0;
 	} else {
-		std::cout << "\nremove tmp file failed" << std::endl;
+		LOG(WARNING) << "\nremove tmp file failed" << std::endl;
 		return -1;
 	}
 }
